Refactored game logic, added a short method to the audio manager, fixed player controller
I have now refactored the game logic into another script and added thorough comments to the code. I also added a small method in the audiomanager for Jacob. Lastly I put the playerController back to how it was except for some small changes that needed to stay to work with the gamelogic script
