+++ Assets/Scenes/MovementAndCamera.unity
@@ -21831,6 +21831,7 @@ GameObject:
   - component: {fileID: 1520365515}
   - component: {fileID: 1520365514}
   - component: {fileID: 1520365513}
+  - component: {fileID: 1520365517}
   m_Layer: 0
   m_Name: Player
   m_TagString: Untagged
@@ -21863,10 +21864,8 @@ MonoBehaviour:
   _collisionLayer:
     serializedVersion: 2
     m_Bits: 8
-  _deathGroundTag: DeathGround
-  _winGroundTag: WinGround
-  _deathTimer: 3
-  _winTimer: 5
+  _isDead: 0
+  _hasWon: 0
 --- !u!65 &1520365514
 BoxCollider:
   m_ObjectHideFlags: 0
@@ -21933,6 +21932,22 @@ Transform:
   - {fileID: 356824512}
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 90, z: 0}
+--- !u!114 &1520365517
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1520365512}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 15bd781e355124028b95fd8163534d7e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _deathGroundTag: DeathGround
+  _winGroundTag: WinGround
+  _deathTimer: 3
+  _winTimer: 5
 --- !u!43 &1525833833
 Mesh:
   m_ObjectHideFlags: 0
